Use OSGi service for UI startup handler
Bundle activators should do as little work as possible. Therefore move
the registration of the window listener out of the bundle activation and
instead have OSGi call back when the application has started. That's
later, that improves startup performance and that should avoid race
conditions on the workbench access.
