+++ net.sf.eclipsecs.ui/META-INF/MANIFEST.MF
@@ -10,7 +10,9 @@ Require-Bundle: net.sf.eclipsecs.core,
  org.eclipse.core.expressions,
  org.eclipse.ui.workbench,
  org.eclipse.jface,
- org.eclipse.help;bundle-version="3.9.0"
+ org.eclipse.help;bundle-version="3.9.0",
+ org.eclipse.e4.ui.workbench;bundle-version="1.13.200"
+Service-Component: OSGI-INF/net.sf.eclipsecs.ui.ApplicationStartedHandler.xml
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Export-Package: net.sf.eclipsecs.ui,
  net.sf.eclipsecs.ui.properties.filter,
@@ -66,6 +68,7 @@ Import-Package: org.apache.commons.lang3;version="3.12.0",
  org.eclipse.ui.texteditor,
  org.eclipse.ui.views.markers,
  org.osgi.framework,
+ org.osgi.service.event;version="[1.4.0,2.0.0)",
  org.osgi.service.prefs
 Automatic-Module-Name: net.sf.eclipsecs.ui
